+++ paddle_billing/Resources/Reports/Operations/__init__.py
@@ -5,5 +5,6 @@ from paddle_billing.Resources.Reports.Operations.CreatePayoutReconciliationRepor
     CreatePayoutReconciliationReport,
 )
 from paddle_billing.Resources.Reports.Operations.CreateProductsAndPricesReport import CreateProductsAndPricesReport
+from paddle_billing.Resources.Reports.Operations.CreateReport import CreateReport
 from paddle_billing.Resources.Reports.Operations.CreateTransactionsReport import CreateTransactionsReport
 from paddle_billing.Resources.Reports.Operations.ListReports import ListReports