+++ paddle_billing/Notifications/Entities/Shared/TaxMode.py
@@ -5,3 +5,4 @@ class TaxMode(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
     AccountSetting: "TaxMode" = "account_setting"
     External: "TaxMode" = "external"
     Internal: "TaxMode" = "internal"
+    Location: "TaxMode" = "location"