+++ paddle_billing/Entities/Shared/SouthKoreaLocalCardType.py
@@ -1,0 +1,27 @@
+from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
+
+
+class SouthKoreaLocalCardType(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
+    BC: "SouthKoreaLocalCardType" = "bc"
+    Citi: "SouthKoreaLocalCardType" = "citi"
+    Hana: "SouthKoreaLocalCardType" = "hana"
+    Hyundai: "SouthKoreaLocalCardType" = "hyundai"
+    Jeju: "SouthKoreaLocalCardType" = "jeju"
+    Jeonbuk: "SouthKoreaLocalCardType" = "jeonbuk"
+    KakaoBank: "SouthKoreaLocalCardType" = "kakaobank"
+    KBank: "SouthKoreaLocalCardType" = "kbank"
+    KDBBank: "SouthKoreaLocalCardType" = "kdbbank"
+    Kookmin: "SouthKoreaLocalCardType" = "kookmin"
+    Kwangju: "SouthKoreaLocalCardType" = "kwangju"
+    Lotte: "SouthKoreaLocalCardType" = "lotte"
+    MG: "SouthKoreaLocalCardType" = "mg"
+    NH: "SouthKoreaLocalCardType" = "nh"
+    Post: "SouthKoreaLocalCardType" = "post"
+    Samsung: "SouthKoreaLocalCardType" = "samsung"
+    SavingsBank: "SouthKoreaLocalCardType" = "savingsbank"
+    Shinhan: "SouthKoreaLocalCardType" = "shinhan"
+    Shinhyup: "SouthKoreaLocalCardType" = "shinhyup"
+    Suhyup: "SouthKoreaLocalCardType" = "suhyup"
+    TossBank: "SouthKoreaLocalCardType" = "tossbank"
+    Unknown: "SouthKoreaLocalCardType" = "unknown"
+    Woori: "SouthKoreaLocalCardType" = "woori"