+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 
 - Added support for `wechat_pay` payment method type. See [related changelog](https://developer.paddle.com/changelog/2025/wechat-pay-payment-method?utm_source=dx&utm_medium=paddle-python-sdk)
 
+### Deprecated
+- `CreateBalanceReport` is deprecated, use `CreatePayoutReconciliationReport` instead, see [changelog](https://developer.paddle.com/changelog/2025/payout-reconciliation-report?utm_source=dx&utm_medium=paddle-python-sdk)
+
 ## 1.12.0 - 2025-11-12
 
 ### Added