+++ CHANGELOG.md
@@ -14,6 +14,9 @@ Check our main [developer changelog](https://developer.paddle.com/?utm_source=dx
 - Added `api_key_exposure.created` event support. See [related changelog](https://developer.paddle.com/changelog/2025/secret-scanning?utm_source=dx&utm_medium=paddle-python-sdk)
 - Added `grand_total_tax` to transaction totals and adjusted totals. See [related changelog](https://developer.paddle.com/changelog/2026/grand-total-tax-field?utm_source=dx&utm_medium=paddle-python-sdk)
 
+### Deprecated
+- `CreateBalanceReport` is deprecated, use `CreatePayoutReconciliationReport` instead, see [changelog](https://developer.paddle.com/changelog/2025/payout-reconciliation-report?utm_source=dx&utm_medium=paddle-python-sdk)
+
 ## 1.12.0 - 2025-11-12
 
 ### Added